+++ src/rider/main/kotlin/fr/socolin/rider/plugins/hsf/models/HsfIconManager.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.IconLoader
 import com.intellij.openapi.util.SystemInfo
+import com.intellij.ui.icons.CachedImageIcon
 import com.intellij.ui.scale.ScaleContext
 import com.intellij.ui.scale.UserScaleContext
 import com.intellij.util.io.systemIndependentPath
@@ -119,71 +120,13 @@ class HsfIconManager(private val project: Project) {
 
     private fun loadIconFromDisk(iconFile: Path): Icon? {
         try {
-            val imageDataByUrlLoaderClass = Class.forName("com.intellij.openapi.util.ImageDataByUrlLoader")
-            val imageDataByUrlLoaderConstructor =
-                imageDataByUrlLoaderClass.constructors.find { c -> c.parameterCount == 4 }
-            val filePathUrl =
-                if (SystemInfo.isWindows) "file:/" + iconFile.systemIndependentPath else "file://" + iconFile.systemIndependentPath
-            val resolver = imageDataByUrlLoaderConstructor!!.newInstance(
-                URL(filePathUrl),
-                iconFile.systemIndependentPath,
-                null,
-                false
-            );
-            val resolveMethod = imageDataByUrlLoaderClass.getDeclaredMethod("resolve");
-            resolveMethod.invoke(resolver);
-            val cachedImageIconClass = Class.forName("com.intellij.openapi.util.CachedImageIcon")
-            val scalableIconClass = Class.forName("com.intellij.openapi.util.ScalableIcon")
-            val scaleContextAwareClass = Class.forName("com.intellij.ui.scale.ScaleContextAware")
-            val cachedImageIconConstructor =
-                cachedImageIconClass.constructors.find { c -> c.parameterCount == 2 && c.parameterTypes[0].name == "java.lang.String" }
-            var cachedImageIcon = cachedImageIconConstructor!!.newInstance(
-                iconFile.systemIndependentPath, resolver
-            )
-            val scaleToWidth = scalableIconClass.getDeclaredMethod("scaleToWidth", Float::class.java);
-            val copyMethod = cachedImageIconClass.getDeclaredMethod("copy");
-            val getScaleContextMethod = scaleContextAwareClass.getDeclaredMethod("getScaleContext");
-            val updateScaleContextMethod =
-                scaleContextAwareClass.getDeclaredMethod("updateScaleContext", UserScaleContext::class.java);
-
-            val scaleContext = ScaleContext.create()
-            if (getScaleContextMethod.invoke(cachedImageIcon) != scaleContext) {
-                // honor scale context as 'iconCache' doesn't do that
-                cachedImageIcon = copyMethod(cachedImageIcon);
-                updateScaleContextMethod.invoke(cachedImageIcon, scaleContext)
-            }
-            return scaleToWidth.invoke(cachedImageIcon, JBUI.pixScale(16.0f)) as Icon
+            val filePathUrl = if (SystemInfo.isWindows) "file:/" + iconFile.systemIndependentPath else "file://" + iconFile.systemIndependentPath
+            return IconLoader.findIcon(URL(filePathUrl));
         } catch (e: Exception) {
             logger.error("Failed to load project icon $iconFile", e)
         }
         return null;
     }
-    /*    private fun loadIconFromDiskInternalApi(iconFile: Path): Icon? {
-            try {
-                val filePathUrl = if (SystemInfo.isWindows)  "file:/" + iconFile.systemIndependentPath else  "file://" + iconFile.systemIndependentPath
-                val resolver =
-                    ImageDataByUrlLoader(
-                        URL(filePathUrl),
-                        iconFile.systemIndependentPath,
-                        null,
-                        false
-                    )
-                resolver.resolve()
-                var cachedImageIcon = CachedImageIcon(iconFile.systemIndependentPath, resolver)
-
-                val scaleContext = ScaleContext.create()
-                if (cachedImageIcon.scaleContext != scaleContext) {
-                    // honor scale context as 'iconCache' doesn't do that
-                    cachedImageIcon = cachedImageIcon.copy()
-                    cachedImageIcon.updateScaleContext(scaleContext)
-                }
-
-                return cachedImageIcon.scaleToWidth(JBUI.pixScale(16.0f))
-            } catch (e: Exception) {
-                logger.error("Failed to load project icon $iconFile", e)
-            }
-            return null;
-        }*/
 
     private fun addIcon(hsfIcon: HsfIcon) {
         allIcons.add(hsfIcon)